// src/api/generated/models/leagueResource.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */
import type { OwnerResource } from './ownerResource.ts';
import type { SubscriptionResource } from './subscriptionResource.ts';

export interface LeagueResource {
  id: string;
  name: string;
  /** @nullable */
  silver_limit: number | null;
  /** @nullable */
  golden_limit: number | null;
  /** @nullable */
  black_limit: number | null;
  mulct_contract_limit: number;
  /** @nullable */
  player_limit: number | null;
  subscription_id: string;
  subscription_start: string;
  subscription_end: string;
  is_active: boolean;
  owner?: OwnerResource;
  subscription?: SubscriptionResource;
  created_at: string;
  updated_at: string;
}